refactor: merge constants.ts into lib/constants/ barrel
Move DUE_DAY_OPTIONS and DUE_DAY_LABELS into lease.ts.
Create index.ts that re-exports all submodules so
existing @/lib/constants imports continue to work.
Delete old constants.ts.
