--- resources/js/lib/constants.ts
@@ -1,19 +1,0 @@
-export const DUE_DAY_OPTIONS = [
-    { value: '1', label: '1st' },
-    { value: '5', label: '5th' },
-    { value: '10', label: '10th' },
-    { value: '15', label: '15th' },
-    { value: '20', label: '20th' },
-    { value: '25', label: '25th' },
-    { value: '31', label: 'Last day of month' },
-];
-
-export const DUE_DAY_LABELS: Record<number, string> = {
-    1: '1st',
-    5: '5th',
-    10: '10th',
-    15: '15th',
-    20: '20th',
-    25: '25th',
-    31: 'Last day of month',
-};
